backend/requirements.txt
# Core
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0

# Database
mysqlclient==2.2.0
PyMySQL==1.1.0

# Security
bcrypt==4.1.2
python-dotenv==1.0.0

# Validation
marshmallow==3.20.1
flask-marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3

# Development
pytest==7.4.3
pytest-flask==1.3.0
black==23.12.1
flake8==7.0.0
